fix lint by organizing errors to a file for simplifications
