+++ server/services/moderationConfigService/moderationConfigService.ts
@@ -4,12 +4,7 @@ import { type JsonObject, type ReadonlyDeep } from 'type-fest';
 
 import { type ConsumerDirectives } from '../../lib/cache/index.js';
 import type { Invoker } from '../../models/types/permissioning.js';
-import {
-  CoopError,
-  ErrorType,
-  type ErrorInstanceData,
-} from '../../utils/errors.js';
-import { __throw } from '../../utils/misc.js';
+import { type RuleErrorType, type LocationBankErrorType } from './errors.js';
 import { type ModerationConfigServicePg } from './dbTypes.js';
 import { type Action, type Policy } from './index.js';
 import ActionOperations, {
@@ -454,50 +449,3 @@ export class ModerationConfigService implements ReturnsModerationConfigTypes {
   }
 }
 
-type RuleErrorType =
-  | 'RuleNameExistsError'
-  | 'RuleHasRunningBacktestsError'
-  | 'RuleIsMissingContentTypeError';
-
-// TODO: throw this error as appropriate on failed rule creation/update.
-export const makeRuleNameExistsError = (data: ErrorInstanceData) =>
-  new CoopError({
-    status: 409,
-    type: [ErrorType.UniqueViolation],
-    title: 'A rule with that name already exists in this organization.',
-    name: 'RuleNameExistsError',
-    ...data,
-  });
-
-// TODO: throw this error as appropriate on failed rule creation/update.
-export const makeRuleIsMissingContentTypeError = (data: ErrorInstanceData) =>
-  new CoopError({
-    status: 400,
-    type: [ErrorType.InvalidUserInput],
-    title: 'This rule must contain a content type on which to operate.',
-    name: 'RuleIsMissingContentTypeError',
-    ...data,
-  });
-
-// TODO: throw this error as appropriate on failed rule creation/update.
-export const makeRuleHasRunningBacktestsError = (data: ErrorInstanceData) =>
-  new CoopError({
-    status: 409,
-    type: [ErrorType.AttemptingToMutateActiveRule],
-    title:
-      "This rule cannot be updated while it has running backtests, which are using the rule's current conditions.",
-    name: 'RuleHasRunningBacktestsError',
-    ...data,
-  });
-
-type LocationBankErrorType = 'LocationBankNameExistsError';
-
-// TODO: throw this error as appropriate on failed bank creation/update.
-export const makeLocationBankNameExistsError = (data: ErrorInstanceData) =>
-  new CoopError({
-    status: 409,
-    type: [ErrorType.UniqueViolation],
-    title: 'A location bank with this name already exists',
-    name: 'LocationBankNameExistsError',
-    ...data,
-  });
